+++ components/TextControls/tests/unit/Theming/MDCFilledTextFieldThemingTests.m
@@ -24,7 +24,7 @@ static const CGFloat kFilledSublayerFillColorNormalOpacity = (CGFloat)0.12;
 static const CGFloat kTextColorNormalOpacity = (CGFloat)0.87;
 static const CGFloat kNormalLabelColorNormalOpacity = (CGFloat)0.60;
 
-static const CGFloat kPrimaryAssistiveLabelColorNormalOpacity = 0.60;
+static const CGFloat kPrimaryAssistiveLabelColorNormalOpacity = (CGFloat)0.60;
 static const CGFloat kPrimaryFloatingLabelColorNormalOpacity = (CGFloat)0.60;
 static const CGFloat kPrimaryFloatingLabelColorEditingOpacity = (CGFloat)0.87;
 static const CGFloat kPrimaryUnderlineColorNormalOpacity = (CGFloat)0.42;
